Class "jaspList"

This class represents a list of strings, ints, doubles or booleans/logicals. It is not used directly but exists mainly as a part of jaspTable. It can be indexed through a fieldname and/or a numerical index.

Methods

add(element)

Adds 'element' to the end of the indexable list.

[[fieldName]]:

Retrieves whatever is stored under 'fieldName' in the mapping.

[[fieldName]] <- element:

Stores 'element' under 'fieldName' in the mapping.

[[numericalIndex]]:

Retrieves whatever is stored under 'numericalIndex' in the contiguous list.

[[numericalIndex]] <- element:

Stores 'element' in the contiguous list under index 'numericalIndex'.

Extends

Class "jaspObject"

Examples

# Adding a title
table$colNames[[0]] <- "niceName"

#> Error in `*tmp*`$colNames: object of type 'closure' is not subsettable

table$colTitles[["niceName"]] <- "Gorgeous Title"

#> Error in `*tmp*`$colTitles: object of type 'closure' is not subsettable