jaspOutputObjR

Super object of jaspObjects that show output. Not to be used directly.

Super class

jaspResults::jaspObjR -> jaspOutputObjR

Active bindings

position

the position of a jaspObject, accepts numeric input

title

the title of a jaspObject, accepts string input

info

the info of a jaspObject, accepts string input

Methods

Public methods

jaspOutputObjR$new()

jaspOutputObjR$printHtml()

jaspOutputObjR$setError()

jaspOutputObjR$getError()

jaspOutputObjR$addCitation()

Inherited methods

jaspResults::jaspObjR$dependOn()

jaspResults::jaspObjR$print()

Method new()

Do not use.

Usage

jaspOutputObjR$new()

Method printHtml()

print jaspObject as html.

Usage

jaspOutputObjR$printHtml()

Method setError()

set an error on this jaspObject. If this is a container then the message `x` will be shown on the first child that is a plot or table.

Usage

jaspOutputObjR$setError(x)

Arguments

x,

string, error message.

Method getError()

check whether a jaspObject has an error or not, returns a boolean.

Usage

jaspOutputObjR$getError()

Method addCitation()

set an citation this jaspObject. If this is a container then the citation `x` will also be set on all children.

Usage

jaspOutputObjR$addCitation(x)

Arguments

x,

string, citation.