jaspPlot

A jaspObject to show plots

createJaspPlot(
  plot = NULL,
  title = "",
  width = 320,
  height = 320,
  aspectRatio = 0,
  error = NULL,
  dependencies = NULL,
  position = NULL
)

Arguments

plot
either an object that can be drawn with plot (e.g., a ggplot2 or qgraph object) or a function that draws a plot (e.g., for base graphics)

title
String, title of the jaspPlot

width
Numeric, width of the figure

height
Numeric, height of the figure

aspectRatio,
Numeric, the ratio of width to height

dependencies
see [`jaspOutputObjR$position()`][jaspOutputObjR]

position
see [`jaspOutputObjR$position()`][jaspOutputObjR]

info
String.

jaspObject
Do not use.

Super classes

jaspResults::jaspObjR -> jaspResults::jaspOutputObjR -> jaspPlotR

Public fields

aspectRatio,

numeric, the ratio of width to height

Active bindings

plotObject

R object or NULL, the plotObject

aspectRatio,

numeric, the ratio of width to height

width

of the figure

height

of the figure

status

of the jaspObject

Methods

Public methods

jaspPlotR$new()

Inherited methods

jaspResults::jaspObjR$dependOn()

jaspResults::jaspObjR$print()

jaspResults::jaspOutputObjR$addCitation()

jaspResults::jaspOutputObjR$getError()

jaspResults::jaspOutputObjR$printHtml()

jaspResults::jaspOutputObjR$setError()

Method new()

Create a new jaspPlot object, however, createJaspPlot is preferred.

Usage

jaspPlotR$new(
  plot = NULL,
  title = "",
  width = 320,
  height = 320,
  aspectRatio = 0,
  error = NULL,
  dependencies = NULL,
  position = NULL,
  info = NULL,
  jaspObject = NULL
)

Arguments

plot

either an object that can be drawn with plot (e.g., a ggplot2 or qgraph object) or a function that draws a plot (e.g., for base graphics)

title

String, title of the jaspPlot

width

Numeric, width of the figure

height

Numeric, height of the figure

aspectRatio,

Numeric, the ratio of width to height

dependencies

see [`jaspOutputObjR$position()`][jaspOutputObjR]

position

see [`jaspOutputObjR$position()`][jaspOutputObjR]

info

String.

jaspObject

Do not use.