Easy results for your JASP analysis

This package helps you build your analysis for JASP (http://jasp-stats.org). Using the included classes jaspResults, jaspTable and others you can easily structure output into tables, collections and plots. When you then later on run this in JASP it will automatically make it appear fully formatted in the resultsview. When developing your analysis in R(-Studio) you can see a simpler representation by using $print().

Details

See the following pages for more information on how to use this package: jaspObject, jaspContainer, jaspTable, jaspPlot, jaspHtml, jaspList and of course jaspResultsClass.

Author

Joris Goosen j.c.goosen@uva.nl

Maintainer: JASP Team <info@jasp-stats.org>