Class "jaspResultsClass"

This class is the base of the jaspResults object that is used to store all results from an analysis and which, in JASP, sends data to the resultsview.

Creation

initJaspResults():

When you are running your code seperately from JASP you can call initJaspResults() to have 'jaspResults' be created. If you are making an analysis in JASP you do not need to do anything as jaspResults will be a given to your analysis.

Methods

length():

Returns how many objects are stored in this container. (inherited from jaspContainer)

[[:

Retrieve an object from this container as specified under the fieldname. (inherited from jaspContainer)

[[<-:

Insert an object into this container under a field-name, if this object is a jaspObject and without a title it will get the field-name as title. (inherited from jaspContainer)

setOptions:

Uses the passed list as current options. This is usually passed by JASP directly but could be useful to set when developing your analysis, aka in jaspTools.

changeOptions:

Changes the currently set options and removes all objects that depend on the changed options. Mostly useful for unit tests because this we we can simulate re-running the analysis. Should not be used in an analysis!

getKeepList:

Internal function, generates a list of all files that should be kept, aka plot-images, statefile and save location of the results themselves.

getPlotObjectsForState:

Internal function.

setErrorMessage:

Sets an errormessage on the results, should not be called from an analysis. You can set errors on plots and tables directly.

complete:

Calls send and changes status to complete and stores itself next to the state. This should not be called by a user, JASP will take care of it,

send:

Send results to JASP but obviously only when running the analysis *in* JASP.

Fields

relativePathKeep:

Location of the temporary files, only for internal use in JASP.

Extends

Class "jaspContainer"

Examples

#creating a table in the results:
jaspResults[["stats"]] <- createJaspTable("Descriptives")

# Checking if the dependencies removed the object and creating it again if necessary:
if(is.null(jaspResults[["stats"]])) 
  jaspResults[["stats"]] <- makeANiceDescriptivesTableFunction()